move logic to obi_one to remove dependency to app
