+++ obi_one/scientific/library/morphology_template.json
